1. Added whatIf support - fixed issue #2 2. Added logging and quiet support - fixed issue #1 3. Added command line arguments parsing

+++ bin/kuduSync.js
@@ -147,46 +147,55 @@ var Manifest = (function () {
     };
     return Manifest;
 })();
-function kuduSync(fromPath, toPath, nextManifestPath, previousManifestPath) {
+function kuduSync(fromPath, toPath, nextManifestPath, previousManifestPath, whatIf) {
     Ensure.argNotNull(fromPath, "fromPath");
     Ensure.argNotNull(toPath, "toPath");
     Ensure.argNotNull(nextManifestPath, "nextManifestPath");
     var from = new DirectoryInfo(fromPath);
     var to = new DirectoryInfo(toPath);
     var nextManifest = new Manifest();
-    kuduSyncDirectory(from, to, from.path(), to.path(), Manifest.load(previousManifestPath), nextManifest);
-    Manifest.save(nextManifest, nextManifestPath);
+    log("Kudu sync from: " + from.path() + " to: " + to.path());
+    kuduSyncDirectory(from, to, from.path(), to.path(), Manifest.load(previousManifestPath), nextManifest, whatIf);
+    if(!whatIf) {
+        Manifest.save(nextManifest, nextManifestPath);
+    }
 }
 exports.kuduSync = kuduSync;
-function copyFile(fromFile, toFilePath) {
+function copyFile(fromFile, toFilePath, whatIf) {
     Ensure.argNotNull(fromFile, "fromFile");
     Ensure.argNotNull(toFilePath, "toFilePath");
     log("Copy file from: " + fromFile.path() + " to: " + toFilePath);
-    fs.createReadStream(fromFile.path()).pipe(fs.createWriteStream(toFilePath));
+    if(!whatIf) {
+        fs.createReadStream(fromFile.path()).pipe(fs.createWriteStream(toFilePath));
+    }
 }
-function deleteFile(file) {
+function deleteFile(file, whatIf) {
     Ensure.argNotNull(file, "file");
     var path = file.path();
     log("Deleting file: " + path);
-    fs.unlinkSync(path);
+    if(!whatIf) {
+        fs.unlinkSync(path);
+    }
 }
-function deleteDirectoryRecursive(directory) {
+function deleteDirectoryRecursive(directory, whatIf) {
     Ensure.argNotNull(directory, "directory");
     var path = directory.path();
     log("Deleting directory: " + path);
     var files = directory.files();
     for(var fileKey in files) {
         var file = files[fileKey];
-        deleteFile(file);
+        deleteFile(file, whatIf);
     }
     var subDirectories = directory.subDirectories();
     for(var subDirectoryKey in subDirectories) {
         var subDirectory = subDirectories[subDirectoryKey];
-        deleteDirectoryRecursive(subDirectory);
+        deleteDirectoryRecursive(subDirectory, whatIf);
     }
-    fs.rmdirSync(path);
+    if(!whatIf) {
+        fs.rmdirSync(path);
+    }
 }
-function kuduSyncDirectory(from, to, fromRootPath, toRootPath, manifest, outManifest) {
+function kuduSyncDirectory(from, to, fromRootPath, toRootPath, manifest, outManifest, whatIf) {
     Ensure.argNotNull(from, "from");
     Ensure.argNotNull(to, "to");
     Ensure.argNotNull(fromRootPath, "fromRootPath");
@@ -196,15 +205,16 @@ function kuduSyncDirectory(from, to, fromRootPath, toRootPath, manifest, outMani
     if(from.isSourceControl()) {
         return;
     }
-    log("Copy directory from: " + from.path() + " to: " + to.path());
-    to.ensureCreated();
+    if(!whatIf) {
+        to.ensureCreated();
+    }
     var fromFiles = from.files();
-    var toFiles = to.files();
+    var toFiles = getFilesConsiderWhatIf(to, whatIf);
     for(var toFileKey in toFiles) {
         var toFile = toFiles[toFileKey];
         if(!fromFiles[toFile.name()]) {
             if(manifest.isEmpty() || manifest.isPathInManifest(toFile.path(), toRootPath)) {
-                deleteFile(toFile);
+                deleteFile(toFile, whatIf);
             }
         }
     }
@@ -213,16 +223,16 @@ function kuduSyncDirectory(from, to, fromRootPath, toRootPath, manifest, outMani
         outManifest.addFileToManifest(fromFile.path(), fromRootPath);
         var toFile = toFiles[fromFile.name()];
         if(toFile == null || fromFile.modifiedTime() > toFile.modifiedTime()) {
-            copyFile(fromFile, pathUtil.join(to.path(), fromFile.name()));
+            copyFile(fromFile, pathUtil.join(to.path(), fromFile.name()), whatIf);
         }
     }
     var fromSubDirectories = from.subDirectories();
-    var toSubDirectories = to.subDirectories();
+    var toSubDirectories = getSubDirectoriesConsiderWhatIf(to, whatIf);
     for(var toSubDirectoryKey in toSubDirectories) {
         var toSubDirectory = toSubDirectories[toSubDirectoryKey];
         if(!fromSubDirectories[toSubDirectory.name()]) {
             if(manifest.isEmpty() || manifest.isPathInManifest(toSubDirectory.path(), toRootPath)) {
-                deleteDirectoryRecursive(toSubDirectory);
+                deleteDirectoryRecursive(toSubDirectory, whatIf);
             }
         }
     }
@@ -230,22 +240,44 @@ function kuduSyncDirectory(from, to, fromRootPath, toRootPath, manifest, outMani
         var fromSubDirectory = fromSubDirectories[fromSubDirectoryKey];
         outManifest.addFileToManifest(fromSubDirectory.path(), fromRootPath);
         var toSubDirectory = new DirectoryInfo(pathUtil.join(to.path(), fromSubDirectory.name()));
-        kuduSyncDirectory(fromSubDirectory, toSubDirectory, fromRootPath, toRootPath, manifest, outManifest);
+        kuduSyncDirectory(fromSubDirectory, toSubDirectory, fromRootPath, toRootPath, manifest, outManifest, whatIf);
     }
 }
-try  {
-    if(process.argv.length < 5) {
-        console.log("Usage: kuduSync [from directory path] [to directory path] [next manifest file path] [previous manifest file path (optional)]");
-    } else {
-        var from = process.argv[2];
-        var to = process.argv[3];
-        var nextManifestPath = process.argv[4];
-        var previousManifestPath = null;
-        if(process.argv.length > 5) {
-            var previousManifestPath = process.argv[5];
+function getFilesConsiderWhatIf(dir, whatIf) {
+    try  {
+        return dir.files();
+    } catch (e) {
+        if(whatIf) {
+            return [];
         }
-        kuduSync(from, to, nextManifestPath, previousManifestPath);
+        throw e;
     }
+}
+function getSubDirectoriesConsiderWhatIf(dir, whatIf) {
+    try  {
+        return dir.subDirectories();
+    } catch (e) {
+        if(whatIf) {
+            return [];
+        }
+        throw e;
+    }
+}
+var commander = require("commander");
+commander.version("0.0.1").option("-f, --fromDir [dir path]", "Source directory to sync (* required)").option("-t, --toDir [dir path]", "Destination directory to sync (* required)").option("-p, --previousManifest [manifest file path]", "Previous manifest file path (* required)").option("-n, --nextManifest [manifest file path]", "Next manifest file path (optional)").option("-q, --quiet", "No logging").option("-w, --whatIf", "Only log without actual copy/remove of files").parse(process.argv);
+try  {
+    var commanderValues = commander;
+    var fromDir = commanderValues.fromDir;
+    var toDir = commanderValues.toDir;
+    var previousManifest = commanderValues.previousManifest;
+    var nextManifest = commanderValues.nextManifest;
+    var quiet = commanderValues.quiet;
+    var whatIf = commanderValues.whatIf;
+    if(quiet) {
+        log = function () {
+        };
+    }
+    kuduSync(fromDir, toDir, nextManifest, previousManifest, whatIf);
 } catch (e) {
-    log("Error: " + e);
+    log("" + e);
 }
